+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Metadata object has new methods for working with unlisted attribute
 - Metadata method deprecated= now accepts and return true/false
 - Included date_facts in building a manifest
+- Using API for user provisioning by login
 - Added method for updating report definition in easy way - GoodData::Report#update_definiton(opts, &block)
 - GoodData::MdObject.json is assignable now (using attr_accessor)